+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/acl/TrackedEntityProgramOwnerService.java
@@ -25,21 +25,44 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.hisp.dhis.trackedentity;
+package org.hisp.dhis.tracker.acl;
 
-import java.util.List;
-import java.util.Set;
-import org.hisp.dhis.common.GenericStore;
+import org.hisp.dhis.organisationunit.OrganisationUnit;
 import org.hisp.dhis.program.Program;
+import org.hisp.dhis.trackedentity.TrackedEntity;
+import org.hisp.dhis.trackedentity.TrackedEntityProgramOwner;
 
 /**
  * @author Ameen Mohamed
  */
-public interface TrackedEntityProgramOwnerStore extends GenericStore<TrackedEntityProgramOwner> {
-  String ID = TrackedEntityProgramOwnerStore.class.getName();
+public interface TrackedEntityProgramOwnerService {
+  String ID = TrackedEntityProgramOwnerService.class.getName();
 
-  /** Get tracked entity program owner entity for the te-program combination. */
+  /**
+   * Get the program owner details for a tracked entity.
+   *
+   * @return The TrackedEntityProgramOwner object
+   */
   TrackedEntityProgramOwner getTrackedEntityProgramOwner(TrackedEntity te, Program program);
 
-  List<TrackedEntityProgramOwnerOrgUnit> getTrackedEntityProgramOwnerOrgUnits(Set<Long> teIds);
+  /**
+   * Assign an orgUnit as the owner for a tracked entity for the given program. If another owner
+   * already exist then it would be overwritten.
+   */
+  void createOrUpdateTrackedEntityProgramOwner(
+      TrackedEntity trackedEntity, Program program, OrganisationUnit orgUnit);
+
+  /**
+   * Update the owner ou for a tracked entity for the given program. If no owner previously exist,
+   * then this method will fail.
+   */
+  void updateTrackedEntityProgramOwner(
+      TrackedEntity trackedEntity, Program program, OrganisationUnit orgUnit);
+
+  /**
+   * Create a new program owner ou for a tracked entity. If an owner previously exist, then this
+   * method will fail.
+   */
+  void createTrackedEntityProgramOwner(
+      TrackedEntity trackedEntity, Program program, OrganisationUnit orgUnit);
 }
